[K8s-Gen] Encapsulate ADT in objects
