--- src/main/scala/temple/generate/kube/ast/OrchestrationRoot.scala
@@ -1,6 +1,0 @@
-package temple.generate.kube.ast
-
-/** Input information to generate kubernetes scripts */
-case class OrchestrationRoot(services: Seq[Service])
-
-case class Service(name: String, image: String, ports: Seq[Int])
